Initial Custom material support
This is still rough outline of how the new custom material system will,
look like, but some basic materials already works (Copper, Porcelain and Glass).

This change also partially removes the dynamic object system, some
parts are still in use but will be moved later. The effect system,
which also uses the dynamic object system, is mostly commented out but
kept as we might want to resurrect it at a later point.