Add support for renderMode of View3D
The default of rendering to a texture while nice for Qt Quick, is not so
performant on embedded devices where rendering to offscreen targets can
be expensive.  So for those cases, there is no a renderMode setting to
allow for the end user to specify how the view is rendered.

Texture Mode: Render scene to FBO and be a texture provider.  All Qt
Quick composition cases work, but at the const of performance and anti-
aliasing options.

RenderNode Mode: Render the scene inline, potentially clobering the 2.5d
Qt Quick renderer.

Underlay Mode: Render the scene before Qt Quick. It is possible to
render 2D content on top of the scene, but not under.

Overlay Mode: Render the scene after Qt Quick. It is possible to render
2D content under the scene, but not over.