Add SkyBox background mode
Implemented by sampling the light probe texture onto a quad that fills the
entire screen, which uses an equirectangular projection.

In the future we might want to consider using separate textures, and/or
different projections. There are some minor issues with texture gradients and
equirectangular projections. These have been worked around by setting the
gradients to zero, which is not optimal, but works OK in practice. If we switch
to using texture cubes instead of equirectangular projections, these issues can
be avoided.