Avoiding copying QDemonRef instances as much as possible
Only when the caller needs to hold onto a type for an extended period of
time should a copy be made.

Note: Due to an over-optimistic approach in the beginning there are some
changes here that don't give us anything, but they are left as-is as it's
too much effort to revert all. The initial approach was to fragile, so I
opted for a more careful approach to avoid having hard to find issues
in the future. There are also places where we could be a lot smarter,
e.g., if we access a ref'd member of a class often without holding onto
it, consider exposing it as a public member or add an operator to access
it without causing unnecessary ref counting.