WasdController: use input handlers
Eat our own dog food (follow our own recommendations), and use
input handlers rather than MouseArea.

The downside is that WasdController cannot be a child inside the
DemonView, but must be made as a sibling for a yet unknown reason.
The upshot is that the application can combine WasdController with
other kinds of input, like a separate tap handler for shooting.

In addition:
    - No reason to have a view property. The WasdController can
        be made a sibling to the DemonView (or eventually, a child),
        or the parent property can be set directly.
    - status object can be a QtObject rather than an item.