QQuick3DNode: add functions to calculate globalTransform in the front-end
In many cases there is no time to wait for the scene graph to update before updating
the global matrix for a node. If you e.g change the position of a node, you
need to be able to e.g query to globalPosition right after in order to
position other objects in the scene relative to it. Waiting for the scenegraph
to update will cause such calculations to end up using old values.