QSSGRenderNode: don't flip z when the orientation is already right-handed
This calculateLocalTransform() creates a right-handed transform.
This means that if the nodes orientation is already right-handed, we
don't need to do any flipping. This is only needed when the
node is left-handed (since we create a right-handed transform). And
then we do it in one go at the end of the function.

If we flip the z value for position/pivot when the orientation is
right-handed, we effectively see the position as left-handed. E.g
if the user sets the z position to be -100 for a right-handed
node, we would flip it to be at position 100 instead. This would
be against the reason the set a right-handed system on a node
in the first place.